backend/feature/user/self (#42)
* feat: dict serialize for User model

* refactor: UserSelfView

* feat: get userself

* feat: UserSelfView
