+++ backend/moneyFUS/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 from .views.user.index import UserView
 from .views.user.login import LoginView
+from .views.user.self import UserSelfView
 from .views.asset.index import AssetView
 from .views.asset.category import UsageCategoryView
 from .views.loan.lend import LendView
@@ -10,6 +11,7 @@ from .views.loan.borrowed import BorrowedView
 urlpatterns = [
     path("user/login", LoginView.as_view()),
     path("user", UserView.as_view()),
+    path("user/self", UserSelfView.as_view()),
     path("asset", AssetView.as_view()),
     path("asset/category", UsageCategoryView.as_view()),
     path("lend", LendView.as_view()),
