+++ backend/moneyFUS/views/user/index.py
@@ -39,33 +39,3 @@ class UserView(APIView):
             {"success": False, "reason": "Invalid credentials"},
             status=status.HTTP_400_BAD_REQUEST,
         )
-
-    def delete(self, request):
-        user = request.user
-        user.is_active = False
-        user.save()
-        res = Response({"success": True}, status=status.HTTP_204_NO_CONTENT)
-        res.delete_cookie("token")
-        return res
-
-    def patch(self, request):
-        user = request.user
-        user.email = request.data.get("email")
-        user.username = request.data.get("username")
-        old_plain_password = request.data.get("old_password")
-        new_plain_password = request.data.get("new_password")
-
-        res = Response()
-
-        try:
-            user.save()
-            res = Response({"success": True})
-        except ValidationError:
-            res = Response({"sucess": False})
-            return res
-
-        if new_plain_password and user.check_password(old_plain_password):
-            user.set_password(new_plain_password)
-            res = Response({"success": True})
-
-        return res
